requirements.txt
Django==1.10.2
django-crispy-forms==1.6.1
django-vanilla-views==1.0.4
djangorestframework==3.5.1
